Handle problems in pr and add test cases
