+++ tests/omni_lock_rust/src/lib.rs
@@ -1,5 +1,6 @@
+pub mod ckb_sys_call;
 pub mod debug_utils;
 pub mod dummy_data_loader;
 pub mod omni_lock;
+pub mod opentx;
 pub mod xudt_rce_mol;
-pub mod ckb_sys_call;
